backend/requirements.txt
alembic==1.9.3
aniso8601==9.0.1
black==23.1.0
blinker==1.5
cachelib==0.10.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
flake8==6.0.0
Flask==2.0
Flask-API==2.0
Flask-Cors==3.0.10
Flask-HTTPAuth==4.5.0
Flask-JWT-Extended
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-OAuthlib==0.9.6
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==2.1.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.1.0
six==1.16.0
SQLAlchemy==2.0.3
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
